restructure dialog calls and menu loop under utils
